Update DLL.cpp
One condition in moveup and one condition in movedown do not work. i marked them with comments. It is when it is moved from the start of the list to the end in moveup and when it is moved from the end of the list to the start of the list in movedown. Also ~DLL does not completely work and neither does random function.